+++ mayan/apps/dynamic_search/icons.py
@@ -30,4 +30,10 @@ icon_search_backend_reindex = Icon(
 icon_search_form_clear = Icon(
     driver_name='fontawesome', symbol='times-circle'
 )
+icon_search_model_detail = Icon(
+    driver_name='fontawesome', symbol='square-binary'
+)
+icon_search_model_list = Icon(
+    driver_name='fontawesome', symbol='file-invoice'
+)
 icon_search_submit = Icon(driver_name='fontawesome', symbol='search')